.github/scripts/requirements.txt
lxml
pyyaml
beautifulsoup4
